Add debug option to assembler.

+++ src/assembler.cpp
@@ -8,7 +8,7 @@
 #include <zconf.h>
 #include <libgen.h>
 
-void Florence::Assembler::assemble(const std::string &input_filename, const std::string &output_filename) {
+void Florence::Assembler::assemble(const std::string &input_filename, const std::string &output_filename, bool is_debug) {
     std::string source = Florence::Utils::read_source(input_filename); // Get the source code.
     auto *logger = new Florence::FSCompiler::ErrorLogger(); // Initialise the error logger.
     auto *tokenizer = new Florence::FSCompiler::Tokenizer(source, logger); // Initialise the tokenizer.
@@ -20,5 +20,7 @@ void Florence::Assembler::assemble(const std::string &input_filename, const std:
     std::string generated_assembly = Florence::Linker::link(source_code); // Link the source code to standard library.
     Florence::Utils::write_source("temp.asm", generated_assembly); // Write the code to an output file
     system(("fasm temp.asm " + output_filename).c_str()); // Create executable.
-    system("rm temp.asm"); // Remove the temporary file created by the process.
+    if (!is_debug) {
+		system("rm temp.asm"); // Remove the temporary file created by the process.
+	}
 }
